Fix for crashes reported in  catsout/wallpaper-engine-kde-plugin#412
The first cause of crashes was that the initial extent was being initialized with dimensions higher than the maximum, whereas the existing code assumed that it would be initialized with width 0. Added handling for this scenario.

The second cause of crashes was that the MDL parser wasn't detecting any bones for some of the puppet warps. This was causing serious issues with Vulkan later on, so I've detected this case and removed affected puppet warps. This prevents the last of the crashing, but it leaves the affected components in essentially a "t-pose" that looks pretty bad. Will need to see why the MDL parser can't find the bones.