Fixed MDL parsing issues related to catsout/wallpaper-engine-kde-plugin#412
It seems the MDLs in this wallpaper use an alternative format that violate some of the assumptions in the existing parser. Added code to detect and handle this "Alternative MDL Format". This allows the wallpaper to render mostly correctly, thought there are still small mismatches where the arms meet the shoulders and the neck meets the body. Will need to investigate further.