Fixed animation issues related to catsout/wallpaper-engine-kde-plugin#412
I found that the way that blending was being done caused a situation where  objects with multiple layers of animations where the total blend value between them was >1 would end up not playing some of the animations. The way the blend was distributed between them was that each one would subtract from the global blend until there was none left for the rest. Since this calculation was done in reverse order, it means that with multiple animation layers with a blend value of 1 each, only the last animation actually worked.

First I changed the way the blend was distributed between animation layers when the sum of the blend values is > 1 by precalculating the sum of the blend values and then dividing each one by that amount. Guaranteeing the total equals 1, but without zeroing out any of the layers.

This got all the animations playing, but for objects with total blend > 1, the resulting animations appeared significantly suppressed, not moving as far as they were supposed to or not rotating as much as they were supposed to. As a result, the mismatch was still there. Simply increasing the distributed blend values would cause the initial positions of the objects to be way off, without really increasing the range of the animation. Ultimately I broke down the animations into the "base_frame" and "deltas". The base frame was the starting point of the animation and so I apply the distributed blend value to that. The deltas are the difference between each frame and the base frame and to these values I apply the original un-distributed blend value of the layer. This appears to correct the issue without any adverse effects on the animations that were working previously.