#6 Fixed buggy animations when using MDLA version 3
After the previous changes I noticed a couple wallpapers that had been working reasonably well before but now had glitchy animations. (e.g. https://steamcommunity.com/sharedfiles/filedetails/?id=2861113973 ).

As it turns out, these wallpapers were previously failing to parse some of their animations, and the previous changes in this branch fixed that, but the animations were still not being read correctly, which ended with a worse result.

I found that some special additional handling is required to correctly parse the space between animations in MDLA version 3. This fixed the wallpapers that were showing regressions, and also fixed the animation issues (but not the color issues) in https://steamcommunity.com/sharedfiles/filedetails/?id=2924727184 which was one of our wallpapers that was previously crashing.

Upon further investigation, I found that with this change, a sizeable chunk of the wallpapers I had previously noted for bugged animations were fixed. Here are the ones I noticed:
https://steamcommunity.com/sharedfiles/filedetails/?id=2998506007
https://steamcommunity.com/sharedfiles/filedetails/?id=2985969668
https://steamcommunity.com/sharedfiles/filedetails/?id=2996693620
https://steamcommunity.com/sharedfiles/filedetails/?id=2995570708
https://steamcommunity.com/sharedfiles/filedetails/?id=2880919159
https://steamcommunity.com/sharedfiles/filedetails/?id=2973502027
https://steamcommunity.com/sharedfiles/filedetails/?id=2967709854c
https://steamcommunity.com/sharedfiles/filedetails/?id=2919907582
https://steamcommunity.com/sharedfiles/filedetails/?id=2920495410
https://steamcommunity.com/sharedfiles/filedetails/?id=2896438701
https://steamcommunity.com/sharedfiles/filedetails/?id=2950008303
https://steamcommunity.com/sharedfiles/filedetails/?id=2990210169